Add files via upload
This Number Guessing Game in this computer select a number which don't know you have guess that number for that you have 10 chance.  There is they help also means you selected number in near to that number is or not you can know by message that this number is low means smaller than a number you select and if high than your selected number is greater than a computer number. And if you select right number so there is message that congratulation you win. This is short discription of my code.